+++ app/models/character/bonus.rb
@@ -5,5 +5,7 @@ class Character
     self.table_name = :character_bonus
 
     belongs_to :bonusable, polymorphic: true
+
+    scope :enabled, -> { where(enabled: true) }
   end
 end