+++ db/migrate/20251223090643_add_enabled_to_character_bonuses.rb
@@ -1,0 +1,5 @@
+class AddEnabledToCharacterBonuses < ActiveRecord::Migration[8.1]
+  def change
+    add_column :character_bonus, :enabled, :boolean, null: false, default: true
+  end
+end