+++ spec/controllers/frontend/characters/bonuses_controller_spec.rb
@@ -53,7 +53,7 @@ describe Frontend::Characters::BonusesController do
 
           expect(response).to have_http_status :ok
           expect(response.parsed_body['bonuses'].size).to eq 1
-          expect(response_values.keys).to contain_exactly('id', 'value', 'dynamic_value', 'comment')
+          expect(response_values.keys).to contain_exactly('id', 'value', 'dynamic_value', 'comment', 'enabled')
         end
       end
     end